+++ src/main/kotlin/edu/ics499/team1/app/security/config/WebMvcConfiguration.kt
@@ -1,18 +1,20 @@
 package edu.ics499.team1.app.security.config
 
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.servlet.config.annotation.CorsRegistry
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
-
-/* Allows all network traffic. We will probably want to scope this down and possible implement
-   per control CORS configuration.
-   See this link for more info: https://www.baeldung.com/spring-cors */
-
-@Configuration
-class WebConfig : WebMvcConfigurer {
-    override fun addCorsMappings(registry: CorsRegistry) {
-        registry.addMapping("/**")
-            .allowedOrigins("http://localhost:5173")
-            .allowedMethods("GET", "POST", "PATCH", "DELETE")
-    }
-}
+//import org.springframework.context.annotation.Configuration
+//import org.springframework.web.servlet.config.annotation.CorsRegistry
+//import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
+//
+///* Allows all network traffic. We will probably want to scope this down and possible implement
+//   per control CORS configuration.
+//   See this link for more info: https://www.baeldung.com/spring-cors */
+//
+//@Configuration
+//class WebConfig : WebMvcConfigurer {
+//    override fun addCorsMappings(registry: CorsRegistry) {
+//        println("addCorsMappings method is being called!")
+//        registry.addMapping("/**")
+//            .allowedOrigins("http://localhost:5173")
+//            .allowedMethods("GET", "POST", "PATCH", "DELETE")
+//            .allowedHeaders("*")
+//    }
+//}
